code fixes and update to VS2017

+++ main/mainworker.cpp
@@ -2072,7 +2072,7 @@ void MainWorker::ProcessRXMessage(const CDomoticzHardwareBase *pHardware, const 
 	if (_log.isTraceEnable()) {
 		char  mes[sizeof(tRBUF)*2+2];
 		char * ptmes = mes;
-		for (int i = 0; i < Len; i++) {
+		for (size_t i = 0; i < Len; i++) {
 			sprintf(ptmes,"%02X", pRXCommand[i]);
 			ptmes += 2;
 		}
